backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pymongo==4.6.0
python-jose[cryptography]==3.3.0
websockets==12.0
pydantic==2.5.0
pydantic-settings==2.0.3
web3==6.11.3
eth-account==0.9.0
python-dotenv==1.0.0
dnspython==2.4.2
requests==2.31.0
httpx==0.27.0
aiohttp==3.9.1
python-socketio==5.10.0
python-multipart==0.0.6